requirements.txt
sqlalchemy
pydantic
python-dotenv